增加openapi对响应的描述

+++ src/main/java/com/cool/request/view/tool/RequestParamCacheManager.java
@@ -1,9 +1,11 @@
 package com.cool.request.view.tool;
 
+import com.cool.request.common.cache.CacheStorageService;
 import com.cool.request.common.constant.CoolRequestConfigConstant;
 import com.cool.request.lib.springmvc.RequestCache;
 import com.cool.request.utils.GsonUtils;
-import com.fasterxml.jackson.core.type.TypeReference;
+import com.cool.request.utils.StringUtils;
+import com.intellij.openapi.application.ApplicationManager;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
@@ -11,6 +13,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
+// TODO: 2024/2/26 合并缓存管理
 public class RequestParamCacheManager {
     private static void createIsNotExist() {
         if (!Files.exists(CoolRequestConfigConstant.CONFIG_CONTROLLER_SETTING)) {
@@ -33,16 +36,23 @@ public class RequestParamCacheManager {
     }
 
     public static void removeCache(String id) {
+        if (StringUtils.isEmpty(id)) return;
+        CacheStorageService cacheStorageService = ApplicationManager.getApplication().getService(CacheStorageService.class);
         Path path = Paths.get(CoolRequestConfigConstant.CONFIG_CONTROLLER_SETTING.toString(), id);
         try {
             Files.deleteIfExists(path);
         } catch (IOException ignored) {
         }
+        cacheStorageService.deleteResponseCache(id);
     }
 
     public static void removeAllCache() {
+        CacheStorageService cacheStorageService = ApplicationManager.getApplication().getService(CacheStorageService.class);
+        cacheStorageService.removeAllCache();
+
         Path path = Paths.get(CoolRequestConfigConstant.CONFIG_CONTROLLER_SETTING.toString());
         try {
+            if (Files.notExists(path)) return;
             Files.list(path).forEach(path1 -> {
                 try {
                     Files.deleteIfExists(path1);
